Fix AlertService issues because of refactor
