+++ proxy.conf.json
@@ -1,6 +1,6 @@
 {
   "/api/*": {
-    "target": "http://localhost:8090",
+    "target": "http://localhost:8080",
     "secure": false,
     "changeOrigin": true,
     "logLevel": "debug"
