+++ src/app/alert/alert.module.ts
@@ -1,16 +1,17 @@
-import { NgModule } from '@angular/core';
+import { ModuleWithProviders, NgModule } from '@angular/core';
 import { CommonModule } from '@angular/common';
-import { MatDialogModule, MatSnackBarModule } from '@angular/material';
+import { MatButtonModule, MatDialogModule, MatSnackBarModule } from '@angular/material';
 import { ServicesModule } from '../services/services.module';
+import { AlertService } from './alert-service/alert.service';
 import { AlertComponent } from './alert.component';
 import { DialogComponent } from './dialog/dialog.component';
 
 @NgModule({
   imports: [
     CommonModule,
-    ServicesModule,
     MatDialogModule,
     MatSnackBarModule,
+    MatButtonModule,
   ],
   declarations: [
     AlertComponent,
@@ -21,7 +22,15 @@ import { DialogComponent } from './dialog/dialog.component';
     AlertComponent,
   ],
   exports: [
-    AlertComponent
+    AlertComponent,
+    DialogComponent,
   ]
 })
-export class AlertModule { }
+export class AlertModule {
+  static forRoot(): ModuleWithProviders {
+    return {
+      ngModule: AlertModule,
+      providers: [AlertService]
+    }
+  }
+}
