+++ src/app/alert/dialog/dialog.component.html
@@ -1,6 +1,6 @@
-<h1 md-dialog-title>Confirm</h1>
-<div id="alert-message" md-dialog-content>{{alert.message}}</div>
-<div md-dialog-actions>
-  <button id="yes-button" md-button (click)="dialogRef.close(true)">Yes</button>
-  <button id="no-button" md-button (click)="dialogRef.close()">No</button>
+<h1 mat-dialog-title>Confirm</h1>
+<div id="alert-message" mat-dialog-content>{{alert.message}}</div>
+<div mat-dialog-actions>
+  <button id="yes-button" mat-button (click)="dialogRef.close(true)">Yes</button>
+  <button id="no-button" mat-button (click)="dialogRef.close()">No</button>
 </div>
