+++ src/app/interviewer/assessment-details/assessment-details.component.html
@@ -15,7 +15,7 @@
 
         <h3>Notes</h3>
         <mat-form-field class="notes-input-container">
-          <textarea mdInput [(ngModel)]="assessment.notes" class="notes-textarea"></textarea>
+          <textarea matInput [(ngModel)]="assessment.notes" class="notes-textarea"></textarea>
         </mat-form-field>
       </mat-card-content>
       <mat-card-actions>
