+++ src/app/interviewer/assessment-list/assessment-list.component.spec.ts
@@ -1,7 +1,7 @@
 import { AssessmentResponse } from '../../domains/assessment-response';
 import { AuthService } from '../../services/auth/auth.service';
 import { Router, ActivatedRoute } from '@angular/router';
-import { AlertService } from '../../services/alert/alert.service';
+import { AlertService } from '../../alert/alert-service/alert.service';
 import { BrowserDynamicTestingModule } from '@angular/platform-browser-dynamic/testing';
 import { BrowserAnimationsModule } from '@angular/platform-browser/animations';
 import { FormsModule, FormBuilder } from '@angular/forms';
