+++ src/app/interviewer/question-dashboard/question-dashboard.component.spec.ts
@@ -1,4 +1,4 @@
-import { AlertService } from '../../services/alert/alert.service';
+import { AlertService } from '../../alert/alert-service/alert.service';
 import { QuestionResponse } from '../../domains/question-response';
 import { AuthService } from '../../services/auth/auth.service';
 import { HttpModule, ResponseOptions, Response } from '@angular/http';
