+++ src/app/public/forgot-password/forgot-password.component.ts
@@ -3,7 +3,7 @@ import { Component, OnDestroy, OnInit, HostBinding } from '@angular/core';
 import { FormBuilder, FormGroup, Validators } from '@angular/forms';
 import { UserService } from '../../services/user/user.service';
 import { ActivatedRoute, Router } from '@angular/router';
-import { AlertService } from '../../services/alert/alert.service';
+import { AlertService } from '../../alert/alert-service/alert.service';
 import { passwordValid, sameValue } from '../../validators';
 import { Subscription } from 'rxjs/Subscription';
 import { ResetForgottenPassword } from 'app/domains/reset-forgotten-password';
