+++ src/app/public/register/register/register.component.ts
@@ -1,5 +1,5 @@
 import { environment } from '../../../../environments/environment';
-import { AlertService } from '../../../services/alert/alert.service';
+import { AlertService } from '../../../alert/alert-service/alert.service';
 import { Router } from '@angular/router';
 import { UserService } from '../../../services/user/user.service';
 import { User } from '../../../domains/user';
